requirements.txt
uvicorn==0.21.1
tensorflow==2.12.0
numpy==1.22.0
Pillow==9.5.0
google-cloud-storage==2.10.0
psutil~=6.0.0
requests~=2.30.0
starlette~=0.26.1
flask~=3.0.3
flask-cors==3.0.10
gunicorn==20.1.0

